Решение с сабмодулем если бы он находился сборщиком, по итогу билда проекта нет из-за отказа билдера добавлять и видеть новый модуль при генерации проекта
